requirements.txt
insightface==0.7.3
opencv-python==4.10.0.84
redis==6.2.0
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.1.1
ffmpeg-python==0.2.0
numpy==1.26.4
pyodbc==5.2.0
onnxruntime
pytz
scikit-image==0.24.0
albumentations
scikit-learn
boxmot
faiss-cpu
websockets
aioredis
motmetrics==1.4.0
pandas
